# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Red Solidaria San Ramón API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns/{id}/donations:
    get:
      operationId: getCampaignDonations
      tags: [donations]
      summary: Get donations for a specific campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of donations for campaign
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Donation"
...
  /donations:
    post:
      operationId: createDonation
      tags: [donations]
      summary: Submit a donation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDonationInput"
      responses:
        "201":
          description: Donation submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
    get:
      operationId: getDonations
      tags: [donations]
      summary: List all donations (admin)
      parameters:
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: List of donations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Donation"
  /donations/{id}:
    get:
      operationId: getDonation
      tags: [donations]
      summary: Get donation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Donation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
    put:
      operationId: updateDonationStatus
      tags: [donations]
      summary: Update donation status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDonationStatusInput"
      responses:
        "200":
          description: Updated donation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
  /donations/{id}/proofs:
    get:
      operationId: getDonationProofs
      tags: [donations]
      summary: List proof images for a donation (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of proofs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DonationProof"
    post:
      operationId: addDonationProof
      tags: [donations]
      summary: Add a manual proof image to a donation (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                imageUrl:
                  type: string
                publicId:
                  type: string
                  nullable: true
                mimeType:
                  type: string
                  nullable: true
              required: [imageUrl]
      responses:
        "201":
          description: Proof added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonationProof"
  /donations/{id}/proofs/{proofId}:
    delete:
      operationId: deleteDonationProof
      tags: [donations]
      summary: Delete a proof image (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: proofId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Donation:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
          nullable: true
        campaignTitle:
          type: string
          nullable: true
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        amount:
          type: number
        paymentMethod:
          type: string
        message:
          type: string
          nullable: true
        anonymous:
          type: boolean
        publicProof:
          type: boolean
        receiptUrl:
          type: string
          nullable: true
        receiptNote:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        adminNote:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - firstName
        - lastName
        - email
        - amount
        - paymentMethod
        - anonymous
        - status
        - createdAt
    CreateDonationInput:
      type: object
      properties:
        campaignId:
          type: integer
          nullable: true
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        amount:
          type: number
        paymentMethod:
          type: string
          enum: [yape, plin, transfer, card, cash, other]
        message:
          type: string
          nullable: true
        anonymous:
          type: boolean
        publicProof:
          type: boolean
        receiptUrl:
          type: string
          nullable: true
        receiptNote:
          type: string
          nullable: true
        proofImageUrl:
          type: string
          nullable: true
        proofPublicId:
          type: string
          nullable: true
        proofMimeType:
          type: string
          nullable: true
      required:
        - firstName
        - lastName
        - email
        - amount
        - paymentMethod
        - anonymous
...
    DonationProof:
      type: object
      properties:
        id:
          type: integer
        donationId:
          type: integer
        imageUrl:
          type: string
        publicId:
          type: string
          nullable: true
        mimeType:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - donationId
        - imageUrl
        - createdAt
...
    UpdateDonationStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected]
        adminNote:
          type: string
          nullable: true
      required: [status]
...