# lib/api-zod/src/generated/types/createDonationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Red Solidaria San Ramón API
 * OpenAPI spec version: 0.2.0
 */
import type { CreateDonationInputPaymentMethod } from "./createDonationInputPaymentMethod";

export interface CreateDonationInput {
  campaignId?: number | null;
  firstName: string;
  lastName: string;
  email: string;
  phone?: string | null;
  amount: number;
  paymentMethod: CreateDonationInputPaymentMethod;
  message?: string | null;
  anonymous: boolean;
  publicProof?: boolean;
  receiptUrl?: string | null;
  receiptNote?: string | null;
  proofImageUrl?: string | null;
  proofPublicId?: string | null;
  proofMimeType?: string | null;
}