# lib/api-zod/src/generated/types/donation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Red Solidaria San Ramón API
 * OpenAPI spec version: 0.2.0
 */
import type { DonationStatus } from "./donationStatus";

export interface Donation {
  id: number;
  campaignId?: number | null;
  campaignTitle?: string | null;
  firstName: string;
  lastName: string;
  email: string;
  phone?: string | null;
  amount: number;
  paymentMethod: string;
  message?: string | null;
  anonymous: boolean;
  publicProof?: boolean;
  receiptUrl?: string | null;
  receiptNote?: string | null;
  status: DonationStatus;
  adminNote?: string | null;
  createdAt: string;
}