# lib/api-zod/src/generated/types/donationProof.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Red Solidaria San Ramón API
 * OpenAPI spec version: 0.2.0
 */

export interface DonationProof {
  id: number;
  donationId: number;
  imageUrl: string;
  publicId?: string | null;
  mimeType?: string | null;
  createdAt: string;
}