# lib/db/src/schema/donations.ts
import { pgTable, text, serial, boolean, real, integer, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const donationsTable = pgTable("donations", {
  id: serial("id").primaryKey(),
  campaignId: integer("campaign_id"),
  firstName: text("first_name").notNull(),
  lastName: text("last_name").notNull(),
  email: text("email").notNull(),
  phone: text("phone"),
  amount: real("amount").notNull(),
  paymentMethod: text("payment_method").notNull(),
  message: text("message"),
  anonymous: boolean("anonymous").notNull().default(false),
  // El donante decide si su comprobante se muestra al público
  publicProof: boolean("public_proof").notNull().default(false),
  receiptUrl: text("receipt_url"),
  receiptNote: text("receipt_note"),
  status: text("status").notNull().default("pending"),
  adminNote: text("admin_note"),
  createdAt: timestamp("created_at").defaultNow().notNull(),
});

export const insertDonationSchema = createInsertSchema(donationsTable).omit({ id: true, createdAt: true });
export type InsertDonation = z.infer<typeof insertDonationSchema>;
export type Donation = typeof donationsTable.$inferSelect;